BE/requirements.txt
alembic==1.15.2
asgiref==3.10.0
blinker==1.9.0
boto3==1.28.62
botocore==1.31.85
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
colorlog==6.7.0
cryptography==41.0.4
Django==5.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.20.1
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
six==1.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.0.7
Werkzeug==2.3.7
